+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors@0.5.0
+  - @volontariapp/errors-nest@0.10.0
+
 ## 2.1.14
 
 ### Patch Changes