+++ packages/domain-event/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.4.10
+
+### Patch Changes
+
+- Updated typeORM version to match database package version
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.10.0
+  - @volontariapp/errors@0.5.0`
+  - @volontariapp/database@1.8.1
+  - @volontariapp/errors-nest@0.9.1
+
 ## 2.4.9
 
 ### Patch Changes